Welcome to chezmoi!
This is the initial commit for my entirely overhauled dotfile
system. Previously, I had been using anishathalye/dotbot for about 15
years.

I think I'm really going to like chezmoi better because it models the
separation between the source of truth and files it creates more
cleanly.

The one thing I think I will miss is that I could simply open a file
in the destination and it would open the version-controled file
instead of having to remember as I'm moving around my system whether
or not a file I'm about to edit is managed by chezmoi.

Oh, this also represents my first move into the wild world of zsh.
There's a lot of change happening. But, I like it.

Signed-off-by: Pepper Lebeck-Jobe <eljobe@gmail.com>