Add my border-shimmer config
This is a config file for the initial version of the eljobe/janky-shimmer
utility I wrote.

It isn't really ready for prime-time. For example, I already know I'm going to
"break" the config file format because I want to be able to make the active and
inactive border styles differ more not only in colors, but also behaviors.